Created typeahead filter.
