+++ app/templates/components/typeahead-filter.hbs
@@ -1,0 +1,1 @@
+<input type="text" class="form-control">
